Add interface for VPN data source
Change VPN profile source to an interface. Preparation to allow managed configurations as a second source
