+++ src/frontends/android/app/src/main/java/org/strongswan/android/ui/VpnProfileListFragment.java
@@ -41,6 +41,7 @@ import android.widget.Toast;
 import org.strongswan.android.R;
 import org.strongswan.android.data.VpnProfile;
 import org.strongswan.android.data.VpnProfileDataSource;
+import org.strongswan.android.data.VpnProfileSource;
 import org.strongswan.android.ui.adapter.VpnProfileAdapter;
 import org.strongswan.android.utils.Constants;
 
@@ -65,17 +66,18 @@ public class VpnProfileListFragment extends Fragment
 	private HashSet<Integer> mSelected;
 	private boolean mReadOnly;
 
-	private BroadcastReceiver mProfilesChanged = new BroadcastReceiver()
+	private final BroadcastReceiver mProfilesChanged = new BroadcastReceiver()
 	{
 		@Override
 		public void onReceive(Context context, Intent intent)
 		{
-			long id, ids[];
+			long id;
+			long[] ids;
 			if ((id = intent.getLongExtra(Constants.VPN_PROFILES_SINGLE, 0)) > 0)
 			{
 				VpnProfile profile = mDataSource.getVpnProfile(id);
 				if (profile != null)
-				{	/* in case this was an edit, we remove it first */
+				{   /* in case this was an edit, we remove it first */
 					mVpnProfiles.remove(profile);
 					mVpnProfiles.add(profile);
 					mListAdapter.notifyDataSetChanged();
@@ -104,7 +106,8 @@ public class VpnProfileListFragment extends Fragment
 	/**
 	 * The activity containing this fragment should implement this interface
 	 */
-	public interface OnVpnProfileSelectedListener {
+	public interface OnVpnProfileSelectedListener
+	{
 		void onVpnProfileSelected(VpnProfile profile);
 	}
 
@@ -119,7 +122,7 @@ public class VpnProfileListFragment extends Fragment
 
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container,
-							 Bundle savedInstanceState)
+	                         Bundle savedInstanceState)
 	{
 		View view = inflater.inflate(R.layout.profile_list_fragment, null);
 
@@ -159,7 +162,7 @@ public class VpnProfileListFragment extends Fragment
 			mSelected = selected != null ? new HashSet<>(selected) : new HashSet<>();
 		}
 
-		mDataSource = new VpnProfileDataSource(this.getActivity());
+		mDataSource = new VpnProfileSource(this.getActivity());
 		mDataSource.open();
 
 		/* cached list of profiles used as backend for the ListView */
@@ -206,19 +209,18 @@ public class VpnProfileListFragment extends Fragment
 	@Override
 	public boolean onOptionsItemSelected(MenuItem item)
 	{
-		switch (item.getItemId())
+		if (item.getItemId() == R.id.add_profile)
 		{
-			case R.id.add_profile:
-				Intent connectionIntent = new Intent(getActivity(),
-													 VpnProfileDetailActivity.class);
-				startActivity(connectionIntent);
-				return true;
-			default:
-				return super.onOptionsItemSelected(item);
+			Intent connectionIntent = new Intent(getActivity(),
+			                                     VpnProfileDetailActivity.class);
+			startActivity(connectionIntent);
+			return true;
 		}
+		return super.onOptionsItemSelected(item);
 	}
 
-	private final OnItemClickListener mVpnProfileClicked = new OnItemClickListener() {
+	private final OnItemClickListener mVpnProfileClicked = new OnItemClickListener()
+	{
 		@Override
 		public void onItemClick(AdapterView<?> a, View v, int position, long id)
 		{
@@ -229,7 +231,8 @@ public class VpnProfileListFragment extends Fragment
 		}
 	};
 
-	private final MultiChoiceModeListener mVpnProfileSelected = new MultiChoiceModeListener() {
+	private final MultiChoiceModeListener mVpnProfileSelected = new MultiChoiceModeListener()
+	{
 		private MenuItem mEditProfile;
 		private MenuItem mCopyProfile;
 
@@ -297,7 +300,7 @@ public class VpnProfileListFragment extends Fragment
 					{
 						profiles.add((VpnProfile)mListView.getItemAtPosition(position));
 					}
-					long ids[] = new long[profiles.size()];
+					long[] ids = new long[profiles.size()];
 					for (int i = 0; i < profiles.size(); i++)
 					{
 						VpnProfile profile = profiles.get(i);
@@ -308,7 +311,7 @@ public class VpnProfileListFragment extends Fragment
 					intent.putExtra(Constants.VPN_PROFILES_MULTIPLE, ids);
 					LocalBroadcastManager.getInstance(getActivity()).sendBroadcast(intent);
 					Toast.makeText(VpnProfileListFragment.this.getActivity(),
-								   R.string.profiles_deleted, Toast.LENGTH_SHORT).show();
+					               R.string.profiles_deleted, Toast.LENGTH_SHORT).show();
 					break;
 				}
 				default:
@@ -320,7 +323,7 @@ public class VpnProfileListFragment extends Fragment
 
 		@Override
 		public void onItemCheckedStateChanged(ActionMode mode, int position,
-											  long id, boolean checked)
+		                                      long id, boolean checked)
 		{
 			if (checked)
 			{
